divvying spaces works
