+++ artificial-intelligence/funcs/battleLogic.js
