+++ artificial-intelligence/funcs/calculateAllotmentPoints.js
