+++ artificial-intelligence/funcs/gridGenerator.js
