+++ artificial-intelligence/funcs/hexHighlighter.js
