+++ artificial-intelligence/funcs/index.js
