+++ artificial-intelligence/funcs/validBoard.js
